Create More options in the game
Extended versions of the game. Now program accept alternative lists of options, like Rock, Paper, Scissors, Lizard, Spock ect.
At this stage, before the start of the game, the user will be able to choose the options that will be used. After entering his/her name, the user should provide a list of options.
Whatever list of options the user chooses, program is able to identify which option beats which, that is, the relationships between different options.